feat: customAlert at workout feedback
